+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added cards offered by Waveshaping to the bottom of the deck.
 - Added cards discovered by Kaldorei Cultivator to the bottom of the deck.
 - Added Timethief Rafaam Counter.
+- Filter the secret given to the opponent by Faceless Enigma.
 - Fixed crashes when Murozond, Unbounded's attack was set to infinity. HDT is now capable of handling all its power.
 - Fixed The Kingslayer not Highlighting Incindius' Eruptions.
 - Fixed fabled bundled cards not appearing as Legendary cards in the decklist.